requirements.txt
# Core ML / DL
torch==2.8.0
torchvision==0.23.0
pandas==2.2.2
numpy==2.0.2
scikit-learn==1.6.1
tqdm==4.67.1

# Experiment tracking & configs
mlflow==3.4.0
hydra-core==1.3.2        # for configs
pyyaml>=6.0               # Hydra uses YAML
python-dotenv>=1.0.0

# Visualization
matplotlib==3.10.0
seaborn>=0.12.2

# Demo
streamlit>=1.24.1
Pillow>=10.0.0 
altair>=5.0.1           